aws/sagemaker/requirements.txt
ultralytics>=8.0.0
opencv-python-headless
pillow
numpy
